try and separate Context classes into own files
