+++ package.json
@@ -73,11 +73,11 @@
   "devDependencies": {
     "cheerio": "^0.18.0",
     "esperanto": "^0.6.17",
-    "gobble": "^0.7.8",
+    "gobble": "^0.9.2",
     "gobble-babel": "^4.0.0",
-    "gobble-cli": "^0.3.5",
+    "gobble-cli": "^0.4.1",
     "gobble-es6-transpiler": "^0.1.2",
-    "gobble-esperanto": "^0.2.3",
+    "gobble-esperanto": "file:../gobble-esperanto",
     "gobble-esperanto-bundle": "^0.1.3",
     "gobble-sorcery": "^0.3.0",
     "jshint": "^2.5.11",
