--- src/viewmodel/models/Index.js
@@ -1,1 +1,0 @@
-/* moved to Context */
