--- src/viewmodel/models/MemberReference.js
@@ -1,1 +1,0 @@
-/* moved to Context */
